+++ packages/Charts/src/tests/radar-viewer-test.ts
@@ -50,6 +50,7 @@ category('Radar', () => {
       showMin: false,
       showMax: false,
       showValues: false,
+      normalization: 'Column',
       valuesColumnNames: ['AGE', 'COUNTRYID'],
       legendVisibility: 'Auto',
     };
@@ -62,6 +63,7 @@ category('Radar', () => {
       showTooltip: false,
       backgroundMaxColor: 0xFF4F616F,
       showValues: false,
+      normalization: 'Global',
     });
 
     const options = await getOptions(changedViewer);
@@ -69,6 +71,7 @@ category('Radar', () => {
     expect(options.showTooltip, false);
     expect(options.backgroundMaxColor, 0xFF4F616F);
     expect(options.showValues, false);
+    expect(options.normalization, 'Global');
   });
 
   after(async () => {